+++ api-derive/democracy/preimages.ts
@@ -2,16 +2,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.8.0';
-import type { AccountId, Balance, BlockNumber, Call, Hash, PreimageStatus } from 'https://deno.land/x/polkadot@0.2.21/types/interfaces/index.ts';
-import type { FrameSupportPreimagesBounded, PalletPreimageRequestStatus } from 'https://deno.land/x/polkadot@0.2.21/types/lookup.ts';
-import type { Bytes, Option } from 'https://deno.land/x/polkadot@0.2.21/types-codec/mod.ts';
-import type { ITuple } from 'https://deno.land/x/polkadot@0.2.21/types-codec/types/index.ts';
-import type { HexString } from 'https://deno.land/x/polkadot@0.2.21/util/types.ts';
+import type { AccountId, Balance, BlockNumber, Call, Hash, PreimageStatus } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { FrameSupportPreimagesBounded, PalletPreimageRequestStatus } from 'https://deno.land/x/polkadot/types/lookup.ts';
+import type { Bytes, Option } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { ITuple } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
 import type { DeriveApi, DeriveProposalImage } from '../types.ts';
 
 import { map, of, switchMap } from 'https://esm.sh/rxjs@7.8.0';
 
-import { BN_ZERO, isFunction } from 'https://deno.land/x/polkadot@0.2.21/util/mod.ts';
+import { BN_ZERO, isFunction } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { firstMemo, memo } from '../util/index.ts';
 import { getImageHashBounded } from './util.ts';
@@ -76,7 +76,9 @@ function parseImage (api: DeriveApi, [proposalHash, status, bytes]: [HexString, 
   return { at: BN_ZERO, balance, proposal, proposalHash, proposer };
 }
 
-function getDemocracyImages (api: DeriveApi, hashes: (Hash | Uint8Array | string)[]): Observable<(DeriveProposalImage | undefined)[]> {
+function getDemocracyImages (api: DeriveApi, bounded: (Hash | Uint8Array | string | FrameSupportPreimagesBounded)[]): Observable<(DeriveProposalImage | undefined)[]> {
+  const hashes = bounded.map((b) => getImageHashBounded(b));
+
   return api.query.democracy.preimages.multi<Option<PreimageStatus>>(hashes).pipe(
     map((images): (DeriveProposalImage | undefined)[] =>
       images.map((imageOpt) => parseDemocracy(api, imageOpt))
@@ -86,6 +88,7 @@ function getDemocracyImages (api: DeriveApi, hashes: (Hash | Uint8Array | string
 
 function getImages (api: DeriveApi, bounded: (FrameSupportPreimagesBounded | Uint8Array | string)[]): Observable<(DeriveProposalImage | undefined)[]> {
   const hashes = bounded.map((b) => getImageHashBounded(b));
+  const bytesType = api.registry.lookup.getTypeDef(api.query.preimage.preimageFor.creator.meta.type.asMap.key).type;
 
   return api.query.preimage.statusFor.multi(hashes).pipe(
     switchMap((optStatus) => {
@@ -93,9 +96,13 @@ function getImages (api: DeriveApi, bounded: (FrameSupportPreimagesBounded | Uin
       const keys = statuses
         .map((s, i) =>
           s
-            ? s.isRequested
-              ? [hashes[i], s.asRequested.len.unwrapOr(0)]
-              : [hashes[i], s.asUnrequested.len]
+            ? bytesType === 'H256'
+              // first generation
+              ? hashes[i]
+              // current generation (H256,u32)
+              : s.isRequested
+                ? [hashes[i], s.asRequested.len.unwrapOr(0)]
+                : [hashes[i], s.asUnrequested.len]
             : null
         )
         .filter((p) => !!p);
@@ -121,7 +128,7 @@ export function preimages (instanceId: string, api: DeriveApi): (hashes: (Hash |
   return memo(instanceId, (hashes: (Hash | Uint8Array | string | FrameSupportPreimagesBounded)[]): Observable<(DeriveProposalImage | undefined)[]> =>
     hashes.length
       ? isFunction(api.query.democracy.preimages)
-        ? getDemocracyImages(api, hashes as string[])
+        ? getDemocracyImages(api, hashes)
         : isFunction(api.query.preimage.preimageFor)
           ? getImages(api, hashes)
           : of([])