+++ types-codec/extended/OptionBool.ts
@@ -1,10 +1,10 @@
 // Copyright 2017-2023 @polkadot/types-codec authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { HexString } from 'https://deno.land/x/polkadot@0.2.21/util/types.ts';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
 import type { AnyBool, Inspect, Registry } from '../types/index.ts';
 
-import { isHex, isU8a, u8aToU8a } from 'https://deno.land/x/polkadot@0.2.21/util/mod.ts';
+import { isHex, isU8a, u8aToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { Option } from '../base/Option.ts';
 import { bool as Bool } from '../native/Bool.ts';