+++ types/interfaces/consensus/types.ts
@@ -1,8 +1,8 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { U8aFixed } from 'https://deno.land/x/polkadot@0.2.21/types-codec/mod.ts';
-import type { AccountId } from 'https://deno.land/x/polkadot@0.2.21/types/interfaces/runtime/index.ts';
+import type { U8aFixed } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { AccountId } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
 
 /** @name AuthorityId */
 export interface AuthorityId extends AccountId {}